% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barnacle.R
\docType{data}
\name{barnacle}
\alias{barnacle}
\title{Barnacle dataset.}
\format{\code{barnacle} is a data frame with 3 variables (columns).
\describe{
 \item{F}{Factor indicating...}
 \item{RC}{Rostro-carinal lenght (in mm).}
 \item{DW}{Dry weight (in g.)}
 }}
\usage{
barnacle
}
\description{
Barnacle dataset.
}
\examples{
data(barnacle)
head(barnacle)
}

