% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/globaltest.R
\name{globaltest}
\alias{globaltest}
\title{Testing the equality of M curves specific to each level}
\usage{
globaltest(formula, data = data, der, weights = NULL, nboot = 200,
  h0 = -1, h = -1, nh = 30, kernel = "epanech", p = 3, kbin = 100,
  seed = NULL)
}
\arguments{
\item{formula}{An object of class \code{formula}: a sympbolic description
of the model to be fitted.}

\item{data}{A data frame or matrix containing the model response variable
and covariates required by the \code{formula}.}

\item{der}{Number which determines any inference process.
By default \code{der} is \code{NULL}. If this term is \code{0},
the testing procedures is applied for the estimate. If it is \code{1} or
\code{2}, it is designed for the first or second derivative, respectively.}

\item{weights}{Prior weights on the data.}

\item{nboot}{Number of bootstrap repeats.}

\item{h0}{The kernel bandwidth smoothing parameter for the global effect.
Large values of bandwidth make smoother estimates, smaller values of
bandwidth make less smooth estimates. The default is a bandwidth compute by
cross validation.}

\item{h}{The kernel bandwidth smoothing parameter for the partial effects.}

\item{nh}{Integer number of equally-spaced bandwidth on which the
\code{h} is discretised, to speed up computation.}

\item{kernel}{Character which determines the smoothing kernel.
By default \code{kernel = "epanech"}, this is, the Epanechnikov
density function. Also, several types of kernel funcitons
can be used:  triangular and Gaussian density function,
with \code{"triang"} and \code{"gaussian"} term, respectively.}

\item{p}{Degree of polynomial used.  Its value must be the value of
derivative + 1. The default value is 3 due to the function
returns the estimation, first and second derivative.}

\item{kbin}{Number of binning nodes over which the function
is to be estimated.}

\item{seed}{Seed to be used in the bootstrap procedure.}
}
\value{
The \eqn{T} value and the \eqn{p}-value  are returned. Additionally,
it is shown the decision, accepted or rejected, of the global test.
The null hypothesis is rejected if the \eqn{p}-value\eqn{< 0.05}.
}
\description{
\code{globaltest}  can be used to test the equality of the
\eqn{M} curves specific to each level.
}
\details{
\code{globaltest}  can be used to test the equality of the \eqn{M}
curves specific to each level. This bootstrap based test assumes the
following null hypothesis

\eqn{H_0: m_1 = \ldots = m_M}

Note that, if \eqn{H_0} is not rejected, then the equality of critical points
will also accepted. To test the null hypothesis, it is used an statistic,
\eqn{T}, based on direct nonparametric estimates of the curves.
If the null hypothesis is true, the \eqn{T} value should be close to zero
but is generally greater. The test rule based on \eqn{T} consists of
rejecting the null hypothesis if \eqn{T > T^{1- \alpha}}, where \eqn{T^p}
is the empirical \eqn{p}-percentile of \eqn{T} under the null hypothesis.
}
\examples{
library(NPRegfast)
data(barnacle)
globaltest(DW ~ RC : F, data = barnacle, der = 1, seed = 130853)
}
\author{
Marta Sestelo, Nora M. Villanueva and Javier Roca-Pardinas.
}

