% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/localtest.R
\name{localtest}
\alias{localtest}
\title{Testing the equality of critical points}
\usage{
localtest(formula, data = data, der, weights = NULL, nboot = 500,
  h0 = -1, h = -1, nh = 30, kernel = "epanech", p = 3, kbin = 100,
  rankl = NULL, ranku = NULL, seed = NULL)
}
\arguments{
\item{formula}{An object of class \code{formula}: a sympbolic description
of the model to be fitted.}

\item{data}{A data frame or matrix containing the model response variable
and covariates required by the \code{formula}.}

\item{der}{Number which determines any inference process.
By default \code{der} is \code{NULL}. If this term is \code{0},
the testing procedures is applied for the estimate. If it is \code{1} or
\code{2}, it is designed for the first or second derivative, respectively.}

\item{weights}{Prior weights on the data.}

\item{nboot}{Number of bootstrap repeats.}

\item{h0}{The kernel bandwidth smoothing parameter for the global effect.
Large values of bandwidth make smoother estimates, smaller values of
bandwidth make less smooth estimates. The default is a bandwidth compute by
cross validation.}

\item{h}{The kernel bandwidth smoothing parameter for the partial effects.}

\item{nh}{Integer number of equally-spaced bandwidth on which the
\code{h} is discretised, to speed up computation.}

\item{kernel}{Character which determines the smoothing kernel.
By default \code{kernel = "epanech"}, this is, the Epanechnikov
density function. Also, several types of kernel funcitons
can be used:  triangular and Gaussian density function,
with \code{"triang"} and \code{"gaussian"} term, respectively.}

\item{p}{Degree of polynomial used.  Its value must be the value of
derivative + 1. The default value is 3 due to the function
returns the estimation, first and second derivative.}

\item{kbin}{Number of binning nodes over which the function
is to be estimated.}

\item{rankl}{Number or vector specifying the minimum value for the
interval at which to search the \code{x} value which maximizes the
estimate, first or second derivative  (for each level). The default
is the minimum data value.}

\item{ranku}{Number or vector specifying the maximum value for the
interval at which to search the \code{x} value which maximizes the
estimate, first or second derivative  (for each level). The default
is the maximum data value.}

\item{seed}{Seed to be used in the bootstrap procedure.}
}
\value{
The estimate of \eqn{d} value is returned and its confidence interval
for a specific-level of confidence, i.e. 95\%. Additionally, it is shown
the decision, accepted or rejected,  of the local test. Based on the null
hypothesis is rejected if a zero value is not within the interval.
}
\description{
This function can be used to test the equality of the
\eqn{M} critical points estimated from the respective level-specific curves.
}
\details{
\code{localtest} can be used to test the equality of the
\eqn{M} critical points estimated from the respective level-specific curves.
Note that, even if the curves and/or their derivatives are different, it is
possible for these points to be equal.

For instance, taking the maxima of the first derivatives into account,
interest lies in testing the following null hypothesis

\deqn{H_0: x_{01} = \ldots = x_{0M}}

versus the general  alternative

\deqn{H_1: x_{0i} \ne x_{0j}  \quad {\rm{for}} \quad {\rm{some}} \quad
\emph{i}, \emph{j} \in \{ 1, \ldots, M\}.}

The above hypothesis is true if \eqn{d=x_{0j}-x_{0k}=0} where
\deqn{ (j,k)= argmax \quad (l,m) \quad \{1 \leq l<m \leq M\} \quad |x_{0l}-x_{0m}|, }

otherwise  \eqn{H_0} is false. It is important to highlight that, in practice,
the true \eqn{x_{0j}} are not known, and consequently neither is \eqn{d},
so an estimate \eqn{\hat d = \hat x_{0j}-\hat x_{0k}} is used, where,
in general, \eqn{\hat x_{0l}} are the estimates of \eqn{x_{0l}} based on the
estimated curves \eqn{\hat m_l} with \eqn{l = 1, \ldots , M}.

Needless to say,
since \eqn{\hat d} is only an estimate of the true \eqn{d}, the sampling
uncertainty of these estimates needs to be acknowledged. Hence, a confidence
interval \eqn{(a,b)} is created for \eqn{d} for a specific level of
confidence (95\%).  Based on this, the null hypothesis is rejected if
zero is not contained in the interval.

Note that if this hypothesis is rejected (and the factor has more than
two levels), one option could be to use the \code{maxp.diff} function in
order to obtain the differences between each pair of factor's levels.
}
\examples{
library(npregfast)
data(barnacle)
localtest(DW ~ RC : F, data = barnacle, der = 1, seed = 130853)
}
\author{
Marta Sestelo, Nora M. Villanueva and Javier Roca-Pardinas.
}
\references{
Sestelo, M. (2013). Development and computational implementation of
estimation and inference methods in flexible regression models.
Applications in Biology, Engineering and Environment. PhD Thesis, Department
of Statistics and O.R. University of Vigo.
}

