% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxp.diff.R
\name{maxp.diff}
\alias{maxp.diff}
\title{Differences between the estimation of maximum points
for two factor's levels}
\usage{
maxp.diff(model, factor1 = NULL, factor2 = NULL, der = NULL)
}
\arguments{
\item{model}{Parametric or nonparametric regression model
obtained by \code{\link{frfast}} function.}

\item{factor1}{First factor's level at which to perform the differences
between maximum points.}

\item{factor2}{Second factor's level at which to perform the differences
between maximum points.}

\item{der}{Number which determines any inference process. By default
\code{der} is \code{NULL}. If this term is \code{0}, the calculate of the
differences for maximum point is for the estimate. If it is \code{1} or
\code{2}, it is designed for the first or second derivative, respectively.}
}
\value{
An object is returned with the following elements:
\item{maxp.diff}{a table with a couple of factor's level where it is used
to calculate the differences between maximum points, and their
95\% interval confidence (for the estimation, first and second derivative).}
}
\description{
Differences between the estimation of \code{\link{maxp}} for two
factor's levels. \code{maxp}, a returned element  of class
\code{\link{frfast}}, is the value of covariate \code{x} which maximizes
the  estimate, first or second derivative.
}
\details{
Differences are calculated by subtracting a factor relative to
another (\eqn{factor2 - factor1}).  By default \code{factor2} and
\code{factor1} are \code{NULL}, so the differences calculated are for all
possible combinations between two factors.
}
\examples{
library(NPRegfast)
data(barnacle)
fit2 <- frfast(DW ~ RC : F, data = barnacle) # with interactions
maxp.diff(fit2)
maxp.diff(fit2, der = 1)
maxp.diff(fit2, factor1="marta",factor2="nora")
}
\author{
Marta Sestelo, Nora M. Villanueva and Javier Roca-Pardinas.
}

