% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxp.R
\name{maxp}
\alias{maxp}
\title{Maximum points for the estimate, first and second derivative,
with their 95\% confidence intervals}
\usage{
maxp(model, der = NULL)
}
\arguments{
\item{model}{Parametric or nonparametric regression out
obtained by \code{\link{frfast}} function.}

\item{der}{Number which determines any inference process. By default
\code{der} is \code{NULL}. If this term is \code{0}, the calculation of the
maximum point is for the estimate. If it is \code{1} or \code{2}, it is
designed for the first or second derivative, respectively.}
}
\value{
An object is returned with the following elements:
\item{Estimation}{outputs for estimation with maximum points and their
95\% confidence intervals (for each level).}
\item{First_der}{outputs for first derivative with maximum points and their
95\% confidence intervals (for each level).}
\item{Second_der}{outputs for second derivative with maximum points and their
95\% confidence intervals (for each level).}
}
\description{
Value of covariate \code{x} which maximizes the  estimate,
first and second derivative, for each level of the factor.
}
\examples{
library(NPRegfast)
data(barnacle)

fit <- frfast(DW ~ RC, data = barnacle) # without interactions
maxp(fit)
maxp(fit, der = 0)
maxp(fit, der = 1)
maxp(fit, der = 2)

fit2 <- frfast(DW ~ RC : F, data = barnacle) # with interactions
maxp(fit2)
maxp(fit2, der = 0)
maxp(fit2, der = 1)
maxp(fit2, der = 2)
}
\author{
Marta Sestelo, Nora M. Villanueva and Javier Roca-Pardinas.
}

