% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.frfast.R
\name{summary.frfast}
\alias{print.frfast}
\alias{summary.frfast}
\title{Summarizing fits of \code{frfast} class}
\usage{
\method{summary}{frfast}(model)
}
\arguments{
\item{model}{a fitted \code{frfast} object as produced by \code{frfast()}.}
}
\value{
\code{summary.frfast} computes and returns a list of summary
information for a fitted \code{frfast} object.
\item{model}{type of estimate.}
\item{h}{the kernel bandwidth smoothing parameter.}
\item{dp}{degree of a polynomial.}
\item{nboot}{number of bootstrap repeats.}
\item{kbin}{number of binning nodes over which the function is to be estimated.}
\item{n}{total number of data.}
\item{fmod}{factor's levels.}
}
\description{
Takes a fitted \code{frfast} object produced by \code{frfast()}
and produces various useful summaries from it.
}
\details{
\code{print.frfast} tries to be smart about \code{summary.frfast}.
}
\examples{
library(NPRegfast)
data(barnacle)

# Nonparametric regression without interactions
fit <- frfast(DW ~ RC, data = barnacle)
fit
summary(fit)

# Nonparametric regression with interactions
fit2 <- frfast(DW ~ RC : F, data = barnacle)
fit2
summary(fit2)
}
\author{
Marta Sestelo, Nora M. Villanueva and Javier Roca-Pardinas.
}

